+++ CHANGES.md
@@ -31,7 +31,7 @@ Changes in 9.0
 - `intl` PHP extension is now required in diagnostic.php & INSTALL.md
 - Fix PHP8.1 deprecated strftime() use strftime_compat() instead in Side.php, Date.php, PHPCompatibility.php, strftime_compat.php, Dashboard.inc.php & Preferences.php
 - Add $course_period_id param to limit check to a single Course Period in Courses.fnc.php & Courses.php
-- Add title to Contact & Address button images in Address.inc.php
+- Add title to Contact & Address button images in Address.inc.php & GetStuList.fnc.php
 - CSS select max-width 440px in stylesheet.css & zresponsive.css
 - HTML add label to Points inputs to correct alignment in Grades.php
 - HTML add a11y-hidden label to select in CategoryBreakdown.php, CategoryBreakdownTime.php & StudentFieldBreakdown.php